feat(auth): create AuthModule with updated configurations and remove old type definition
